requirements.txt
asgiref==3.8.1
Django==5.1.5
python-decouple==3.8
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
